refactor: ScheduleDrawer을 useOverlay에서 제어하도록 개선
