+++ src/pages/Home/components/HomeContainer/view/ScheduleList/ScheduleList.tsx
@@ -9,11 +9,12 @@ import EasyAuthentication from "@components/sign/EasyAuthentication.tsx";
 import { useAppDispatch } from "@redux/hooks.ts";
 import { changeHideBudgetMode } from "@redux/slices/settingSlice.ts";
 import { INIT_PERIOD, INIT_REPEAT } from "@constants/schedule.ts";
+import { useScheduleDrawer } from "@hooks/useScheduleDrawer.tsx";
 
 function ScheduleList() {
   const dispatch = useAppDispatch();
-  const { date, todaySchedules, isPending, isError, openDrawer } =
-    useSchedule(); // redux가 직접 하도록 개선 예정
+  const { date, todaySchedules, isPending, isError } = useSchedule(); // redux가 직접 하도록 개선 예정
+  const { openScheduleDrawer } = useScheduleDrawer();
   const [authenticationPageOpen, setAuthenticationPageOpen] = useState(false);
 
   if (isPending) {
@@ -48,7 +49,7 @@ function ScheduleList() {
     // setBottomDrawerOpen(true); // 수정 drawer는 bottombar의 drawer를 공유할 수 있도록 수정 예정
     if (schedule) {
       const start = moment(schedule.start_date); // getMonthSchedule api 수정 후 제거 예정
-      openDrawer({
+      openScheduleDrawer({
         ...schedule,
         repeat: INIT_REPEAT(start),
         period: INIT_PERIOD(start),
